Update patch set 2
Patch Set 2:

(9 comments)

Could you run tryjob (CQ+1) when sending a code review request?
It helps us to make sure the CL works properly on various board configurations,
and to find an issue if exists earlier so the iteration could be faster.
Thank you for your cooperation in advance.

Patch-set: 2
CC: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>