tast-tests: Remove "disabled" attr from example.ServoEcho.
Enable the example.ServoEcho test so that it will run
informationally. This test is expected to fail when the
"servo" var isn't passed via "tast run" due to a Servo
device being unavailable, but it's still desirable to run it
now to determine how common that situation is in the lab. As
a result, note that the test will always fail on VM
builders.

BUG=chromium:867312,chromium:967901
TEST=none

Change-Id: Ib87c143e5274c6f6e316b5936416eacfe3dabdd2