Port platform_ToolchainOptions to Tast
Whitelists are from both autotest-private and public.

Changes in the Tast version:
- filepath.Walk is used instead of multiple find commands.
- /debug/elf is used instead of the readelf command.
- info from /debug/elf is used instead of the file command.

Things which were not added to the Tast test:
- Autotest allowed setting the rootdir to an alternate root.  The test
is faster, so it's not any better to run on a smaller set of files.
- Autotest allowed passing in a hardfp option.  This was not enabled by
default and has not been running in the lab.  That check has been skipped.
- the Autotest has a check that has been commented out since 2014.  It
was not ported but is still tracked in crbug.com/417912.
- the Autotest has a fixme bug (that was closed years ago) for an issue
on i?86 architectures.  Since alex/mario/zgb are deprecated, that
differentiation is not ported.

BUG=chromium:963332
TEST=compared found static/dynamic files from both the Autotest and Tast.
Compared all non-whitelisted files for each check.  Compared lists of found
files for different properties (e.g. program headers, dyntags).  Ran on
nautilus, blaze, panther, kip, eve, pit, peppy.

CQ-DEPEND=CL:1652468
Change-Id: I8a6b5ef855d53f68c8c197277a302e4db4ce1a9a