[DO NOT COMMIT] Port platform_ToolchainOptions to Tast
(Only one condition is currently implemented)
Looking for feedback on this draft:
- Does the overall restructure look reasonable
- Should I split this into multiple files (it will get quite long with
    the additional conditions/whitelists)
- Autotest allows passing in a few options - should we keep any
    functionality that hasn't been running in the cq?

The Autotest has ~10 separate (rather complicated) find commands that run
across all of /

This Tast test walks the entire tree. Conditions are added with a struct
containing a whitelist and verification function (which is run against every
ELF file during the walk).

BUG=chromium:963332
TEST=TBD

Change-Id: I8a6b5ef855d53f68c8c197277a302e4db4ce1a9a