Update patch set 3
Patch Set 3:

(15 comments)

Thank you for the review.
Some of the logic (looking only for /dev, filtering the update-engine process, come from the original autotest test. If there was a mistake implementing the exact same logic, we need to fix it. It the old test logic was wrong, we should update the tast test in a follow on change.

Patch-set: 3
Reviewer: Gerrit User 1337146 <1337146@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>