Update patch set 1
Patch Set 1:

(1 comment)

> adb-proxy work added another child process

It is not scalable to have to update this test every time we add a child process. Can we make sure we remove any assumptions about the actual number of children ?

Instead we should eventually have tests that exercises the various features implemented by such or such child. If a child is missing or is in an incorrect state, the test exercising the associated features will discover that, without assumptions about the internal process organisation.

Patch-set: 1
Reviewer: Gerrit User 1238096 <1238096@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>