Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > adb-proxy work added another child process
> 
> It is not scalable to have to update this test every time we add a child process. Can we make sure we remove any assumptions about the actual number of children ?

This test was added when ip-helper moved to a resident process. We needed to ensure that it survived arc cycling. So it was very much the point to check specifically for 2 procs. The same argument can be made for adb-proxy since this subprocess has better be (always) up when ARC boots. Removing the assumption about an actual number of child procs defeats the whole purpose of the test. It wouldn't, for example, catch orphaned procs over arc reboots. If we end up adding more processes then sure we can revisit the basic idea of the test. 

> 
> Instead we should eventually have tests that exercises the various features implemented by such or such child. If a child is missing or is in an incorrect state, the test exercising the associated features will discover that, without assumptions about the internal process organisation.

Agreed but that's not the purpose of this particular test.

Patch-set: 1
Reviewer: Gerrit User 1307765 <1307765@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>