Update patch set 4
Patch Set 4:

(4 comments)

Thanks, this CL looks mostly good!


> Some of the logic (looking only for /dev, filtering the update-engine process, come from the original autotest test. If there was a mistake implementing the exact same logic, we need to fix it. It the old test logic was wrong, we should update the tast test in a follow on change.

Yes, I know the original Autotest does that. But I still suggest removing the filtering in this CL.

Generally, if the original Autotest did some redundant things, and we port it verbatimly, the ported test will just pass, and once we promote it to CQ we will feel hesitant to remove redundant things because it's just working. It's difficult to improve a test after we port it.

In fact we did many improvements to existing tests as we port them. For example, we did not port some of shill (networking) tests at all because they were already covered by unit tests. We implemented P2P-updater tests in a significantly different way (using virtual network devices in Linux namespaces) because the original way (implementing virtual network devices with TUN/TAP) was too complex and hard to maintain.

In this case, I don't see any obvious reason to filter non-/dev devices. I might be wrong and there can be a reason of course, but we can see it by trying to remove it at first and monitoring test results. If the test fails, we can add filtering back. If we don't remove the filtering now, we can never know if it is actually redundant or not, and we miss a chance to improve this test forever.

Patch-set: 4
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>