tast: Test for ec ring device
This test will open the cros-ec-ring buffer for reading and for each
sensor in the ring set the ODR of to the max frequency for the sensor
and collect data for 2 seconds. The test will fail if there are
timestamps out of order or there are less than half the number of
expected readings.

BUG=b:132282412
TEST=go test
     tast run ${REMOTE_IP} hardware.SensorRing

Change-Id: I225bd95e26643c09d04406d60e259fa1a4b1391a
Reviewed-on: https://chromium-review.googlesource.com/1645002
Tested-by: Mathew King <mathewk@chromium.org>
Commit-Ready: Mathew King <mathewk@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>