Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Before starting the review, I'd like to make sure the status of crbug.com/968589,
> which is not updated for a week, and the test looks still continuously failing.
> In general, as announced, Tast team is requesting to give a priority to stabilize existing tests and promote them to CQ, before adding a new test.
> 
> Could you share the current situation? If it could, could you take a look at the bug first?

Yes, we are currently working on the bug. The issue is to do with the virtual keyboard (some weird deadlock), not to do with the test. We're aiming for a fix by the end of this week.

I'll update the bug with the current status.

Patch-set: 2
Reviewer: Gerrit User 1197444 <1197444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>