tast-tests: shill_stability: ask Upstart for PID
The logic we use for determining the shill PID is complicated and can be
much more easily done by simply asking Upstart what it thinks the PID
is.

This also fixes a subtle race condition: if a previous job has recently
re-started shill, shill may not have finished fork()/execve() yet, and
so a 'pgrep shill' will not find it yet. We could try to hack around
this by modifying the shill Upstart job to report 'started' later, but
really this is just a fact of life of many Upstart services -- the
'started' state does not necessarily mean everything is fully "up and
running."

BUG=none
TEST=run back to back with, e.g., network.ConfigureServiceForProfile
TEST=hack this job to run shill.Safe{Stop,Start}() first

Change-Id: I40ebbfbbbae19f0254a787614ca693d8c0405fbd