tast-tests: Updating virtual USB printer tests
The new USB printer setup flow in ChromeOS is to automatically
configure the printer when it is connected to the system. This change
updates the USB printer tests to search for the given name of the
configured printer and proceed rather than explicitly configuring the
printer itself.

BUG=chromium:973637
TEST=Passes tast tests on elm device

Change-Id: Ia97362fec40ccbb00b54e57fcef5b9ef5594c12b
Reviewed-on: https://chromium-review.googlesource.com/1650590
Tested-by: David Valleau <valleau@chromium.org>
Commit-Ready: David Valleau <valleau@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>