video.WebRTCDecodePerf: Increase timeout of the WebRTCDecodePerf Tast test.
This CL increases the timeout of the WebRTCDecodePerf Tast test. The currently
configured timeout of 3 minutes is not always enough. On some platforms a
successful run seems to take more then 2m40s, at which time the test gets
aborted to leave enough time for cleanup. Increased the timeout to 4 minutes.

TEST=video.WebRTCDecodePerf on eve

BUG=chromium:973325
Change-Id: Ic8c845bc2d11933e71eac6765b65303da20d3b42