video.EncodeAccelPerf*: Increase timeout of the EncodeAccelPerf* Tast tests.
This CL increases the timeout of the EncodeAccelPerf* Tast tests. The default
timeout of 2 minutes is not always enough. On some platforms a successful run
seems to take more then 1m30s, at which time the tests gets aborted to leave
enough time for cleanup. Increased the timeout to 3 minutes.

TEST=video.EncodeAccelPerf on eve

BUG=chromium:973325
Change-Id: Ica4be679183f7c53ae1c6dba487683c3d566ede8