tast-tests: Fixing Printer.PrintUSB test
This change updates the RunPrintTest function to either explicitly
configure the printer or wait for it to be configured automatically
depending on whether or not a PPD was provided.

There has also been some flakiness regarding whether or not the "For"
and "Title" fields of the PDF produced in a basic USB print are
populated with the actual username and document title. In order to fix
the flakiness this change adds a function which strips these fields so
that the comparison between the resulting PDF and the golden file will
pass.

BUG=chromium:973637
TEST=All printer.* tests pass on Elm device

Change-Id: I32bbd72d2f91f9bcc18fc7e18427a99c10f17bc6
Reviewed-on: https://chromium-review.googlesource.com/1659008
Tested-by: David Valleau <valleau@chromium.org>
Commit-Ready: David Valleau <valleau@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>