platform.ChromeMlocked: don't fail on all errors
According to the linked bug, we'll sometimes see ESRCH when trying to
get information about processes. It's unclear to me where exactly this
comes from, but it's likely that these errors shouldn't immediately
translate into test failures here anyway. This test is ultimately
looking for a single process that meets certain requirements; if we fail
to parse the data from a process, as long as it's not the process we're
looking for, it doesn't make much sense for this test to give up
entirely.

BUG=chromium:973736
TEST=platform.ChromeMlocked on falco

Change-Id: I627e44ee46b9736c98a62586a2f2ad380faa4f52
Reviewed-on: https://chromium-review.googlesource.com/1659336
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Ready: George Burgess <gbiv@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>