tast-tests: Skip hwsec.ChallengeResponseMount w/o TPM
Add a dependency on the "tpm" feature into the
hwsec.ChallengeResponseMount, so that the test is skipped when
running on a system without the TPM chip functionality. The
functionality that the test verifies - the challenge-response
mount - heavily relies on the TPM, so the test cannot work
without it.

This change should, in particular, suppress the spurious
failures of this test on the "betty" board (which is a VM
without the TPM emulation).

BUG=chromium:832226
TEST=manual

Change-Id: I9563955bdce087fc459051fb193c5c9c478b49e9
Reviewed-on: https://chromium-review.googlesource.com/1659388
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>