tast-test: activity.Close() stops the activity
Now that ARC sessions that be reused by different tests it is important
that each test does a proper clean up.

Before activity.Close() was only freeing some resources, but the
activity, in case there was still runnig, was not stopped. But this
could cause a problem if the ARC session was reused the another text.

This CL fixes that scenario by callining activity.Stop() at
activity.Close() time.

BUG=chromium:974068
TEST=Run all the tests that call activity.Close(). All of them worked as
     expected. Verified tests: arc.HWOverlayTablet, arc.PIP,
     arc.ResizeActivity, arc.LowMemoryKiller

Change-Id: I8e4ca7e96209c013d8a99d89c47f771a1090edbf