tast-tests: hwsec: ClearOwnership
This is the test entry that just clears the TPM followed by simple
verification of TPM reset.

It is the subset of 'RetakeOwnership`, and it is meant to be used by
developers to verify the functionality or as a push-button solution of
TPM reset.

BUG=b:132657813
TEST=tast run ${DUT_IP} hwsec.ClearOwnership

Change-Id: Ifaccb989ac996a5a5b79ca4db8dd517a9ccc5a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1988052
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>