Promote Tast tests: PlayDecodeAccelUsed.* to critical.
video.PlayDecodeAccelUsed.* are important end-to-end tests
to make sure Chrome can use hardware decoder to decode and play
various video formats (H264, VP8, VP9).

Recent test results for PlayDecodeAccelUsed.* except
video.PlayDecodeAccelUsedMSEVP8 are all green:
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedH264
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedVP8
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedVP9
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedMSEH264
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedMSEVP8
https://tastboard.corp.google.com/test/video.PlayDecodeAccelUsedMSEVP9

For video.PlayDecodeAccelUsedMSEVP8, it is flaky only on
guado-kernelnext. As it is an experimental board which is not tested on
CQ, we shall also add the test to CQ.

Hence, I propose promoting video.PlayDecodeAccelUsed* to tast-CQ.

BUG=chromium:976594
TEST=tast run video.PlayDecodeAccelUsed.*

Change-Id: I32608132efd79bcd2c98668c93e66448ef7eeecf