Tast.camera/video.*: Move waiting for idle CPU out of SetUpBenchmark().
This CL moves waiting for the CPU to become Idle out of the SetUpBenchmark()
function. Testcases are now responsible for calling this function at the time
they see fit.

There are multiple reasons for doing this:
- Tests might need to do additional setup (such as logging in, opening tabs)
  before starting the measurement. It makes more sense to wait until that is
  done before waiting for the CPU to become idle.
- Tests might do multiple runs, performing setup each time. We should wait for
  the CPU to become idle each time, without disabling/restoring the CPU
  frequency each time.
- Previous tests might have failed to perform cleanup, making it difficult for
  the CPU to become idle. Often the setup performed by the test itself at a
  later stage will fix this (e.g. by restarting the UI).

This CL only moves WaitUntilIdle() out of the SetupBenchmark() function. Moving
WaitUntilIdle() to a more approriate place will be done in other CL's.

TEST=camera.Decode/EncodeAccelJPEGPerf

BUG=chromium:938728

Change-Id: Iba8d329bef403fc0aa59a48c10cb42c407562e3f