Update patch set 19
Patch Set 19: Commit-Queue+1

> Patch Set 18:
> 
> Also, I'm going to add a black flash check to this test, so the execution of each case is going to be expensive. So I guess it will be better to test each case once in this test. WDYT?

I understand your point. I'm opposed to reducing this to 1 test iteration, at the moment. With 1 iteration, the original bug was not reproducible. It occurred after a few window state transitions via ADB; I believe that value was 4 iterations. With the fix, the original bug was not reproducible. This means that system server was not able to handle many window state transitions. This is why I increased the value, and tbh, 25 was an arbitrary value. But as you discovered, the fix unintentionally introduced a bug in PIP, which is also bad. Ideally, we should have tests to guard against both of these types of regressions. If the tests you write in Q3 are able to quickly test this bug, I'm ok with revisiting this test and reducing the number of test iterations.

Patch-set: 19
Reviewer: Gerrit User 1192540 <1192540@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1