tast-tests: vm: Factor CrostiniStartEverything into a precondition
Historically we dropped all the tests in CrostiniStartEverything as a
way of avoiding the expensive setup cost. Tast now has a feature for
that (preconditions) so this change makes CrostiniStartEverything use
that.

We model the implementation on arc/pre.go. There are several outstanding
items, notably breaking up the test into its subtests.

BUG=chromium:965398
TEST=tast -verbose run $DUT "vm.CrostiniStartEverything"

Change-Id: I72153a3bdae1b7f804c886b834603931ca9f2832