tast-test: return keyboard top row layout
Add function that returns the correct keyboard top row layout.
This is needed since on some Chromebooks Toggle Zoom is mapped to F3 and
on others on F4.
This CL returns a map with the correct mapping.

This is needed for some ARC tests that need to verify that pressing the
"zoom" key works as expected.

BUG=chromium:978910
TEST=tast -verbose run $DUT example.Keyboard
TEST=Tested on Caroline (used F4), tested on Eve (used F3),
     tested on Eve in Tablet mode (virtual keyboard used F4).

Change-Id: I6b82d17e7591a8202d8c24b87ea570eec9313087