tast: Add local test that tests wifi functionality after suspend/resume.
Prior to this change, tests involving the suspend/resume had to be performed in a wifi-cell with servo.

This test demonstrates how Tast can operate through suspend/resume with a local test and perform wifi operations between multiple suspend/resumes.
Caution was taken in terms of initialization and cleanup to make sure
that the DUT can soft recover through errors.

BUG=None
TEST=WifiResume

Change-Id: I147497899fbc7b8e96023b04d5cc6a19f151aaf4