platform.HeavyMemoryUser: Change the functionality for VMs
Changed the way that VMs are started to align with the CrostiniFiles
test. This will allow the ability to push data files and is more
aligned with typical usage.

Changed RunTest in memory_user.go to only start a VM if the test
includes a VMTask. This will allow tests to take advantage of the
structure and run Chrome and ARC, without requiring the DUT to have
crostini enabled.

BUG=chromium:902164
TEST=Ran platform.HeavyMemoryUser

Change-Id: I35fe0139b61fa745a6c2b49f72949cbd215f1a71
Reviewed-on: https://chromium-review.googlesource.com/1679717
Tested-by: Alexis Savery <asavery@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>