tast-tests: Isolate Chrome log for TestPlay-based tests.
TestPlay-based tests are flaky now. During investigation, we found
Chrome log could be trimmed because it is too large as many tests
already run on an existing Chrome instance. A trimmed log is useless
for debugging.

To obtain a full Chrome log, we decide to spend around 10 seconds for
a test to run it in a standalone Chrome session. By doing so, we can
get a Chrome log that contains only one test.

BUG=chromium:941390
TEST=tast run $ip video.PlayVP8 video.PlayVP9 video.PlayH264 video.PlayDecodeAccelUsed*

Change-Id: I73d3a2805ff6da4012a2c4505a1c4f3dcd19957d