tast-tests: Isolate Chrome log for flaky tests.
When investigating flaky tests, we found Chrome log could be trimmed
because it is too large as many tests already run on an existing Chrome
instance. It makes the log useless if messages around the failure are
trimmed.

We came out an workaround: for flaky tests, e.g. TestPlay() tests,
instead of using an existing Chrome instance satisfying ChromeVideo()
precondition, we restart a Chrome instance (of course the args are the
same we used in ChromeVideo() precondition) in test setup phase, and
"stop ui" in test tear-down phase. By doing so, we can spend around
10 seconds for a test to obtain an isolated log that only runs the test.

Also, we put the setup and tear-down phases in TestPlayIsolate() and
keep the test body, TestPlay(), unchaged. We can switch back to run test
with precondition Chrome instance once flaky reason is found so that we
can save 10 seconds for each test run.

Noted that when specifying "Pre:" attribute in a test configuration,
the test cannot restart or exit the Chrome instance.

BUG=chromium:941390
TEST=tast run $ip video.PlayVP8 video.PlayVP9 video.PlayH264 video.PlayDecodeAccelUsed*

Change-Id: I73d3a2805ff6da4012a2c4505a1c4f3dcd19957d