tast-tests: Isolate Chrome log for flaky tests.
When investigating flaky tests, we found Chrome log could be trimmed
because it is too large as many tests already run on an existing Chrome
instance. It makes the log useless if messages around the failure are
trimmed.

We came out an workaround: for flaky tests, e.g. TestPlay() tests,
instead of using an existing Chrome instance satisfying ChromeVideo()
precondition, we restart a Chrome instance (of course the args are the
same we used in ChromeVideo() precondition) in test setup phase, and
"stop ui" in test tear-down phase. By doing so, we spend around
10 seconds for a test to obtain an isolated log that only runs the test.

Once the test is no longer flaky, we can save the Chrome restart time by
assigning Pre: in test config. It will use existing Chrome instance to
run tests requiring the same pre-condition.

BUG=chromium:941390
TEST=tast run $ip video.PlayVP8 video.PlayVP9 video.PlayH264 video.PlayDecodeAccelUsed*

Change-Id: I73d3a2805ff6da4012a2c4505a1c4f3dcd19957d