tast-tests: vm: Extended the CrostiniPre to allow for an image artifact
To prevent network-based flake, we would like to migrate crostini tests
so that they use a pre-built image artifact where possible, rather than
downloading the image.

We refactor the CrostiniPre so that it can source the image either from
a download or from the data dependency. This is a workaround suggested
by nya@ which still requires us to list the data dependency for every
artifact-based crostini test.

As a result of this change, we make several refactorings to the tests
that involve the artifact:
 - We use the component "crostini" rather than "vm" to distinguish from
   tests for termina, or pluginvm
 - We adopt a parametric style, namely that crostini tests should run
   both via-download and via-build-artifact. In discussion with
   tast-users we are modelling the parameterization after the videos/
   tests.

This change affects both vm.CrostiniStartBasic (which is now
crostini.Sanity*) and vm.CrostiniAudioSanity (which is now
crostini.AudioSanityArtifact).

BUG=chromium:979074,chromium:980048,chromium:981426
TEST=tast -verbose run ${DUT} "crostini.*"

Change-Id: Ie67e8d5abbee5bad8f30784a2ea7b860ec807eae