tast-tests: vm: Extended the CrostiniPre to allow for an image artifact
To prevent network-based flake, we would like to migrate crostini tests
so that they use a pre-built image artifact where possible, rather than
downloading the image.

We refactor the CrostiniPre so that it can source the image either from
a download or from the data dependency. This is a workaround suggested
by nya@ which still requires us to list the data dependency for every
artifact-based crostini test.

BUG=chromium:979074
TEST=tast -verbose run $DUT "vm.CrostiniStart*"

Change-Id: Ie67e8d5abbee5bad8f30784a2ea7b860ec807eae