Update patch set 1
Patch Set 1:

Is this needed urgently for something ?

Can we instead try to do the opposite and get rid of the hardcoded 100.115.92.2:9008 socket address altogether ?

For instance this patch will break when we switch ARCVM to use the current ARC++ networking code since the IP will become 100.115.92.2 for ARCVM.

We should discuss on b/129907372 how to deal with UI Automator. cywang@ mentioned before that ui automator uses adb forward tcp:9008 tcp:9008, which should maybe work on with adb client on the host namespace. This has the advantage of not needing to know the container IP.

Otherwise we might be able to set a DNAT rule in iptables nat table for port 9008 but this requires putting the knowledge of the ARC ip somewhere.

Patch-set: 1
Reviewer: Gerrit User 1238096 <1238096@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>