Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is this needed urgently for something ?

I don't think it's really urgent, but UI Automator is used in a good amount of tests. So, it'll be nice to have.

> Can we instead try to do the opposite and get rid of the hardcoded 100.115.92.2:9008 socket address altogether ?
>
> For instance this patch will break when we switch ARCVM to use the current ARC++ networking code since the IP will become 100.115.92.2 for ARCVM.

What is the plan for this? I think if this will happen soon then I might as well abandon this (as this patch is also blocked by http://crrev.com/c/1637171).

> We should discuss on b/129907372 how to deal with UI Automator. cywang@ mentioned before that ui automator uses adb forward tcp:9008 tcp:9008, which should maybe work on with adb client on the host namespace. This has the advantage of not needing to know the container IP.
> 
> Otherwise we might be able to set a DNAT rule in iptables nat table for port 9008 but this requires putting the knowledge of the ARC ip somewhere.

Sgtm! I don't really know how much work that is, but if it is a lot of work, I think a workaround before that would be nice.

Patch-set: 1
Reviewer: Gerrit User 1346383 <1346383@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>