Update patch set 1
Patch Set 1:

> > For instance this patch will break when we switch ARCVM to use the current ARC++ networking code since the IP will become 100.115.92.2 for ARCVM.
> 
> What is the plan for this? I think if this will happen soon then I might as well abandon this (as this patch is also blocked by http://crrev.com/c/1637171).

This is pending on Garrick's work that is almost done now. It should be a couple more weeks or so.


> I don't think it's really urgent, but UI Automator is used in a good amount of tests. So, it'll be nice to have.

Since this patch is really small, if this let's do you do move forward with other tast tasks, then we can merge this patch when blockers are gone, and revert later when both VM and container uses 100.115.92.2. Unless all of Garrick's work lands first and we are able to switch the VM to the same networking setup as the container.

Patch-set: 1
Reviewer: Gerrit User 1238096 <1238096@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>