tast: Use data-key attribute to tap keys.
Previously, we've used the aria label to determine which keys to
tap for virtual keyboard tests. However, this is problematic
because aria labels may be translated, so it can cause failures
on machines with a non-English UI language.

We switch to using 'data-key' attribute, which is a more stable
identifier for keys. However, it's a bit more verbose.

BUG=chromium:972474
TEST=veyron_minnie

Change-Id: I9e2e4793776d5171fb486d7a9eabeec2653cd645