Update patch set 4
Patch Set 4:

Hi Abe-san, just reviving this, PTAL thanks! I think we were talking about moving common test code into a subpackage and simply having different test cases for informational and critical.

I think that's the best way forward. However, I'm currently blocked because the decoder for English (FST) drops messages when it is loading (see the hack in VirtualKeyboardSuggestions to get around this), whereas Mozc works perfectly fine. We're in the process of discussing what the right behaviour for the decoders should be, so for now, it's difficult to merge the two tests.

My preference would be to land this test first and then when the loading issue is resolved, I'll submit another CL to refactor the two tests to be what we discussed. WDYT?

Patch-set: 4
Reviewer: Gerrit User 1197444 <1197444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>