tast-tests: update platform.Memd to handle multiple margin values
We're going to move to multiple values in the margin file, so instead
of trying to parse and reformat everything, let's just store the whole
content of the margin file and then write it all back in when
finished.

Also update the contact.

BUG=chromium:736538
TEST=run platform.Memd on a DUT with multiple values in the margin
file

Change-Id: I82958df26f436df6528ea68b57480d50cb201a70
Reviewed-on: https://chromium-review.googlesource.com/1684471
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>