tast-test: arc.WindowManagerCUJ works on tablet devices
arc.WindowmanagerCUJ is failing on certain tablet devices because the
activity is being launched while the transition from tablet-to-clamshell
mode is still in progress.

At the moment we don't have a reliable way to "wait for transition to
finish". So, this CL adds a timeout of 5 seconds (research in progress
regarding how to get rid of the delay).

This CL also cleans-up the code a bit, a make it more modular. Needed
for a follow-up CL that will include more tests.

BUG=chromium:980711
TEST=Tested on nocture both in tablet and clamshell modes
    $ tast run $DUT arc.WindowManagerCUJ

Change-Id: I8399e5ef430eb53d44e9c89f9735e466c2e7e754
Reviewed-on: https://chromium-review.googlesource.com/1686074
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Ready: Ricardo Quesada <ricardoq@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>