Update patch set 3
Patch Set 3: Verified+1 Commit-Queue+1

I previously made a change to correct the naming of the "disk" group of telemetry items to be named "memory." Unfortunately I was not aware that we had a Tast test for telemetry at the time. It appears that Tast tests are not run in the CQ so I did not notice anything unusual when I was making that change. This fix is attempting to get the platform.Telemetry test back to passing state on Sarien. Also, nami will be broken because this test tries to start wilco_dtc_supportd, however wilco_dtc_supportd is not a dependency of telem, so I've removed it in order to allow this test to pass on the widest set of boards possible.

Patch-set: 3
Reviewer: Gerrit User 1327553 <1327553@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Label: Verified=+1