Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> Could you describe rationale for chumping this change to add new tests?

Apologies for not leaving a comment here. This CL was part of a chain of important test fixes. I considered removing this CL from the chain and rebasing. But that would cause a lot of work for little gain. As this is a new informational test it won't break anything.

more info about the rationale can be found here: crrev.com/c/1662965

Patch-set: 20