tast: Turn on a display before waiting for audio devices
Some devices use their displays as an internal speaker, like monroe.
When a display is closed, the internal speaker is removed, too. It
causes that we can not find the usable speaker on monroe. To fix the
problem, we need to turn on a display to re-enable a HDMI internal
speaker.

BUG=b:135298456
TEST=tast run ${DUT_IP} audio.ALSAConformance
It can pass on monroe even if a display is closed.

Change-Id: I40c3761b1257d8119529a17d69e7274e004bcf8b