tast-test: update point-of-contact for some arc tests
Removes ricardoq@ from PIP test:
As discussed with edcourtney@ he will take over the test.

Replaces ricardoq@ with cruan@ from ResizeActivity:
As discussed with ruanc@ she will investigate why the test is failing
on some boards. She is the owner of that feature.

Replaces ricardoq@ with takise@ from HWOverlayTablet:
takise@ is the new owner of the HWOverlay bug (http://b/120557146)

BUG=None
TEST=./fast_build.sh -T / -C and compiles fine.

Change-Id: I27ec6e6409a7aeb226fa8a8de9e04367159476e4