Update patch set 12
Patch Set 12: Code-Review-1

(6 comments)

General comments/questions:

(1) Why isn't anyone from the WiFi team on the reviewers list? Especially the one you're putting in the "Contacts" field, who was apparently totally unaware of this until recently.

(2) Are you intentionally de-modularizing everything here? Is that a design pattern for Tast? It's likely true that there are a few too many abstraction layers in Autotest sometimes, but this seems to go the opposite extreme -- stuff that will likely be shared in multiple tests (/sys/class/net/ traversal) is stuffed into a single test, so further tests that want it will either end up duplicating it, or forced to refactor everything -- if they notice it at all.

For (2) - maybe it's fine to just stick it here for now, but we've had this disease fairly generally in a variety of WiFi tests -- that stuff gets open-coded in a variety of tests, in slightly different ways.

Patch-set: 12
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1