tast.video.DecodeAccel*: Specify output folder when running video decoder tests.
This CL passes an output folder to the video_decode_accelerator_tests, so we can
collect test artifacts such as the thumbnails images and video frames.

At this time only the thumbnails image will be dumped to disk when the
FlushAtEndOfStream_RenderThumbnails test fails. This test is only ran on nyan_*
boards and the thumbnails image is about 200kb, so this change will not have
much influence on the size of the produced test artifacts.

TEST=video.DecodeAccelH264New on eve

BUG=chromium:964168

Change-Id: Id2009f9c78be8ce85edefb66b727d611866de915
Reviewed-on: https://chromium-review.googlesource.com/1691881
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Ready: David Staessens <dstaessens@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>