Update patch set 2
Patch Set 2:

How flakey is this test by the way? I suppose that it will always pass if there are no black flashes, but if there are black flashes, how many tries does it take to notice them?

If we could somehow record every frame during the transition that sounds like it could be less flakey to me.

Patch-set: 2