tast-test/graphics: add webrtc_video_playback_delay
This CL adds a new graphics tast to collect WebRTC video
playback metrics, for that, it opens a fake webcam and
establishes a remote peer connection with it; it waits until
this is being played back and starts collecting the histogram
Media.VideoFrameSubmitter for 20 seconds. It then adds the
values to a perf output.

Change-Id: I6015d2bf10d537dd02a71e23dbe4ac7df7c50571
Bug: 953253