tast-tests: Fix always passing security.OpenFDs
(This is a re-upload of Evan's CL, additionally marking the test as
informational.)

A recent change added an upstart restart ui command to the test, which
kills and restarts Chrome. Unfortunately, the test then runs before any
renderer processes have started. It passes because it loops over an
empty list and finds no problems.

Wait a little for processes to show up. Once we do that, suddenly the
test starts failing because it's actually running again. So account for
the new dictionary file that Chrome seems to have open.

In order to avoid having this problem again, fail the test if no processes
are found.

For some reason even with the delay I still don't see any plugin processes.
I don't see any from looking at the ps output myself. Perhaps this part of
the test is totally defunct.

BUG=b:136444774,chromium:992469
TEST=Test passes on eve.

Change-Id: I05b2818010c6000d1b1398511839914578a7024d