tast-test: Add Chrome Settings API
This CL adds a wrapper for a subset of the JavaScript settingsPrivate APIs.
In particular it adds these 2 API:
 - DefaultZoom()
 - SetDefaultZoom()

These two API are needed for an ARC test that has to verify that ARC
changes its zoom factor when Chrome changes it (test coming in a future CL).

BUG=chromium:984375
TEST=tested manually by calling:
     settings.DefaultZoom(ctx, tconn)
     settings.SetDefaultZoom(ctx, tconn zoom)

Change-Id: I1094238819f80cd7241875faa6b73a9c149cd625
Reviewed-on: https://chromium-review.googlesource.com/1703148
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Ready: Ricardo Quesada <ricardoq@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>